Extract refresh interval component
